Almost fully functional VidGrid.user.js
Only one catch, you have to have Dev Tools open with the option set not to cache when Dev Tools is open. 

They're using REACT so their page updates aren't re-rendering the page. So I'm using MutationObserver to watch for the room list change so that the user doesn't have to refresh the page every time the room list updates. It needs a little refining but at this point, it does what I intended for it to do.